Configures PostgreSQL and adds utility functions
Migrates from H2 in-memory database to a PostgreSQL database for persistence.
Adds a Docker Compose file for local PostgreSQL setup.

Introduces a utility class with helper functions like sum, multiply, concatenate, isEqual and a misc function that performs operations based on a string argument.

Adds a basic test to pet service that test a successful case and a exception case, while also improving exception management and adds a blacklist.